chore: clearer separations in data type and new types
