+++ app-bpm/app/src/main/java/com/contextaware/app_bpm/data/model/GeoJsonPoint.kt
@@ -1,0 +1,6 @@
+package com.contextaware.app_bpm.data.model
+
+data class GeoJsonPoint(
+    val type: String,
+    val coordinates: List<Double>
+)
