+++ app-bpm/app/src/main/java/com/contextaware/app_bpm/data/model/ParkingEventType.kt
@@ -1,0 +1,10 @@
+package com.contextaware.app_bpm.data.model
+
+import com.google.gson.annotations.SerializedName
+
+enum class ParkingEventType {
+    @SerializedName("park")
+    PARK,
+    @SerializedName("leave")
+    LEAVE
+}
