+++ app-bpm/app/src/main/java/com/contextaware/app_bpm/data/model/PersonalEvent.kt
@@ -8,16 +8,11 @@ data class PersonalEvent(
     val startTime: String,
     @SerializedName("end_time")
     val endTime: String?,
-    val type: String,
+    val type: ParkingEventType,
     @SerializedName("location_point")
     val locationPoint: GeoJsonPoint,
     @SerializedName("user_id")
     val userId: Int,
     @SerializedName("parking_area_id")
     val parkingAreaId: Int
-)
-
-data class GeoJsonPoint(
-    val type: String,
-    val coordinates: List<Double>
 )
